refactor(identity): extract identity domain leaf
