+++ docs/context/architecture/super-admin.md
@@ -4,7 +4,7 @@ status: shipped
 sources:
   - src/treg/api.py
   - src/treg/routers/admin.py
-  - src/treg/routers/dependencies.py
+  - src/treg/domain/identity/access.py
   - src/treg/config.py
 related:
   - architecture/multi-tenancy.md
@@ -16,7 +16,7 @@ related:
 Everything else is org-scoped; super-admin is the one capability that sees **across all orgs**. It's
 deliberately separate from org roles.
 
-## Authorization (`require_superadmin` in `routers.dependencies`) — hybrid
+## Authorization (`require_superadmin` in `domain.identity.access`) — hybrid
 A caller is a super-admin if EITHER:
 - the presented `X-Treg-Token` equals the env `admin_token` (`get_settings().admin_token`, from
   `TREG_ADMIN_TOKEN`), compared with `hmac.compare_digest` → principal `"env-admin"`; OR
