+++ src/treg/domain/identity/__init__.py
@@ -1,0 +1,1 @@
+"""Caller identity, tokens, roles, and authorization."""
